FillBetweenLines functionality
Initially fillBetween functionality only accepted single number for
series1 and series2 and then it used the fill it with specified color in
between this pair of series.
I was working on a project where i wanted to fill color in between
multiple pairs of series not just a single pair so thats how i made
these changes.

Changes made by me also support the old functionality that is if you
specify a single value other then an array, it would still accept it and
do the fill between the lines.